Update README.md
The dataset that we are going to explore is a combination of the datasets of Top 5 Global Tech Organizations namely AMAZON, MICROSOFT, GOOGLE, APPLE, and FACEBOOK. Each dataset represents the daily price action that includes open, high, low and close of Stock Price. It also gives us an idea of the volume of share transactions per day.

Data time frame		- 1 Year [2019-03-04 to 2020-03-03]
The currency of Price data	- In dollars [$]

Analysis Questions

Q.1.Rank the Company names according to Annual Growth Percentage?

Q.2.Find the monthly trend of these companies based on $ value traded? 

Q.3. Compare the risk of stock by organizations based on price fluctuation?

Q.4. Which company is preferred to be the best to invest according to the investors?